# specs/src/main/resources/openapi.yml
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/customer/products": {
      "get": {
        "tags": [
          "product-customer-controller"
        ],
        "operationId": "getCustomerProduct",
        "parameters": [
          {
            "name": "model",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "product-customer-controller"
        ],
        "operationId": "insertIntoCart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      }
    },
    "/customer/printers": {
      "get": {
        "tags": [
          "printer-customer-controller"
        ],
        "operationId": "getCustomerPrinter",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Printer"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "printer-customer-controller"
        ],
        "operationId": "insertCustomerIntoCart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      }
    },
    "/customer/pcs": {
      "get": {
        "tags": [
          "pc-customer-controller"
        ],
        "operationId": "getCustomerPc",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Pc"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "pc-customer-controller"
        ],
        "operationId": "insertIntoCart_2",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      }
    },
    "/customer/laptops": {
      "get": {
        "tags": [
          "laptop-customer-controller"
        ],
        "operationId": "getCustomerLaptop",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Laptop"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "laptop-customer-controller"
        ],
        "operationId": "insertIntoCart_3",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cart"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Cart"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/products": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "operationId": "getProduct",
        "parameters": [
          {
            "name": "model",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "product-controller"
        ],
        "operationId": "insertIntoProduct",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Product"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        }
      }
    },
    "/admin/printers": {
      "get": {
        "tags": [
          "printer-controller"
        ],
        "operationId": "getPrinter",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Printer"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "printer-controller"
        ],
        "operationId": "insertIntoPrinter",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Printer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Printer"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/laptops": {
      "get": {
        "tags": [
          "laptop-controller"
        ],
        "operationId": "getLaptop",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Laptop"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "laptop-controller"
        ],
        "operationId": "insertIntoLaptop",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Laptop"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Laptop"
                }
              }
            }
          }
        }
      }
    },
...
    "/customer/devices": {
      "get": {
        "tags": [
          "device-customer-controller"
        ],
        "operationId": "getCustomerDevice",
        "parameters": [
          {
            "name": "model",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/products/fullproduct": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "operationId": "getJoinedProducts",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductJoined"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/products/all": {
      "get": {
        "tags": [
          "product-controller"
        ],
        "operationId": "getProducts",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/printers/all": {
      "get": {
        "tags": [
          "printer-controller"
        ],
        "operationId": "getPrinters",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Printer"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/laptops/all": {
      "get": {
        "tags": [
          "laptop-controller"
        ],
        "operationId": "getLaptops",
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Laptop"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/devices": {
      "get": {
        "tags": [
          "device-controller"
        ],
        "operationId": "getDevice",
        "parameters": [
          {
            "name": "model",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Device"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Cart": {
        "required": [
          "model",
          "price",
          "username"
        ],
        "type": "object",
        "properties": {
          "orderId": {
            "type": "integer",
            "format": "int64"
          },
          "model": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int64"
          },
          "type": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "Product": {
        "required": [
          "maker"
        ],
        "type": "object",
        "properties": {
          "maker": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
      "Printer": {
        "required": [
          "color",
          "price",
          "type"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int64"
          },
          "model": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Pc": {
        "required": [
          "cd",
          "hd",
          "price",
          "ram",
          "speed"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int64"
          },
          "model": {
            "type": "string"
          },
          "speed": {
            "type": "integer",
            "format": "int32"
          },
          "ram": {
            "type": "integer",
            "format": "int32"
          },
          "hd": {
            "type": "number",
            "format": "double"
          },
          "cd": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "Laptop": {
        "required": [
          "hd",
          "model",
          "price",
          "ram",
          "screen",
          "speed"
        ],
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int64"
          },
          "model": {
            "type": "string"
          },
          "speed": {
            "type": "integer",
            "format": "int32"
          },
          "ram": {
            "type": "integer",
            "format": "int32"
          },
          "hd": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "screen": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ProductJoined": {
        "type": "object",
        "properties": {
          "maker": {
            "type": "string"
          },
          "model": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        }
      },
      "Device": {
        "type": "object",
        "properties": {
          "model": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  }
}